Move mock DOM event generation to ag-charts-test.
